pyproject.toml
[tool.poetry]
name = "kiosk"
version = "0.1.0"
description = "Sofware for a raspberry pi to work as a custom cash register system"
authors = ["Matthias Duve <morzan1001@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
customtkinter = "^5.2.2"
sqlalchemy = "^2.0.45"
py532lib = { git = "https://github.com/Liam-Deacon/py532lib.git" }
gpiod = { version = "^2.4.0", markers = "sys_platform == 'linux'" }
APScheduler = "^3.11.1"
jinja2 = "^3.1.6"
pillow = "^12.0.0"
sounddevice = "^0.5.3"
soundfile = "^0.13.1"
numpy = "^2.3.5"
scipy = "^1.16.3" 
requests = "^2.32.5" 
psycopg2-binary = "^2.9.11"

[tool.poetry.group.dev.dependencies]
alembic = "^1.17.2" 
pylint = "^4.0.4"
flake8 = "^7.3.0"
black = "^25.12.0"
isort = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.pylint.format]
max-line-length = 100

